requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
attrs==25.3.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
discord.py==2.5.2
frozenlist==1.6.0
google-auth==2.40.1
google-auth-oauthlib==1.2.2
gspread==6.2.0
idna==3.10
multidict==6.4.3
oauthlib==3.2.2
propcache==0.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dotenv==1.1.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
urllib3==2.4.0
yarl==1.20.0
